modular Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
greenlet
